Change wanted-by target of daemon to post-user-session
Delaying daemon startup to later stage makes sure that environment is
properly initialized. For instance, wayland might be unavailable yet if
a daemon is started by `user-session.target`.

Fixes #51

I've removed creation of
~/.config/systemd/user/user-session.target.wants
from spec as systemctl creates missing ${target}.wants if it is missing
and it is not clear why mkdir was added in first place.